Task 17.3

+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/invoice/dao/InvoiceDao.java
@@ -7,5 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 @Repository
-public interface InvoiceDao extends CrudRepository<Invoice, Integer> {
+public interface InvoiceDao extends CrudRepository<Invoice, Long> {
+
+    //Invoice getById(int id);
 }
